# backend/LexBoxApi/openapi/public.yaml
openapi: 3.0.4
info:
  title: Lexbox Public Api
  description: This is the public api for LexBox
  version: 1.0.0
paths:
...
  /api/fwlite-release/download-latest:
    get:
      tags:
        - FwLiteRelease
      parameters:
        - name: edition
          in: query
          schema:
            $ref: '#/components/schemas/FwLiteEdition'
      responses:
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/fwlite-release/latest:
    get:
      tags:
        - FwLiteRelease
      parameters:
        - name: edition
          in: query
          schema:
            $ref: '#/components/schemas/FwLiteEdition'
        - name: appVersion
          in: query
          schema:
            type: string
      responses:
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/FwLiteRelease'
            application/json:
              schema:
                $ref: '#/components/schemas/FwLiteRelease'
            text/json:
              schema:
                $ref: '#/components/schemas/FwLiteRelease'
        default:
          description: ''
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/fwlite-release/should-update:
    get:
      tags:
        - FwLiteRelease
      parameters:
        - name: appVersion
          in: query
          schema:
            type: string
        - name: edition
          in: query
          schema:
            $ref: '#/components/schemas/FwLiteEdition'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ShouldUpdateResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ShouldUpdateResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ShouldUpdateResponse'
...
components:
  schemas:
...
    FwLiteEdition:
      enum:
        - Windows
        - Linux
        - Android
        - iOS
        - Mac
        - WindowsAppInstaller
    FwLiteRelease:
      required:
        - version
        - url
      type: object
      properties:
        version:
          type: string
        url:
          type: string
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          pattern: '^-?(?:0|[1-9]\d*)$'
          anyOf:
            - type: integer
            - type: string
            - enum:
                - null
              nullable: true
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
...
    ShouldUpdateResponse:
      required:
        - release
      type: object
      properties:
        release:
          oneOf:
            - enum:
                - null
              nullable: true
            - $ref: '#/components/schemas/FwLiteRelease'
        update:
          type: boolean
...